+++ src/BusinessLogic/ShippingMethod/Models/ShippingMethod.php
@@ -41,6 +41,8 @@ class ShippingMethod extends Entity
         'isShipToAllCountries',
         'shippingCountries',
         'currency',
+        'fixedPrices',
+        'systemDefaults',
     );
     /**
      * Carrier name.
@@ -151,6 +153,20 @@ class ShippingMethod extends Entity
      * @var string
      */
     protected $currency;
+    /**
+     * Key-value pairs of system info IDs and fixed prices in the default currency
+     * (used in multi-store environments when the service currency does not match the system currency).
+     *
+     * @var array
+     */
+    public $fixedPrices;
+    /**
+     * Key-value pairs of system info IDs and whether they are using default pricing policy
+     * (used in multi-store environments when the service currency does not match the system currency).
+     *
+     * @var array
+     */
+    public $systemDefaults;
 
     /**
      * Transforms raw array data to this entity instance.
@@ -620,5 +636,45 @@ class ShippingMethod extends Entity
     public function setCurrency($currency)
     {
         $this->currency = $currency;
+    }
+
+    /**
+     * Returns fixed prices.
+     *
+     * @return array
+     */
+    public function getFixedPrices()
+    {
+        return $this->fixedPrices;
+    }
+
+    /**
+     * Sets fixed prices.
+     *
+     * @param array $fixedPrices
+     */
+    public function setFixedPrices($fixedPrices)
+    {
+        $this->fixedPrices = $fixedPrices;
+    }
+
+    /**
+     * Returns system defaults.
+     *
+     * @return array
+     */
+    public function getSystemDefaults()
+    {
+        return $this->systemDefaults;
+    }
+
+    /**
+     * Sets system defaults.
+     *
+     * @param array $systemDefaults
+     */
+    public function setSystemDefaults($systemDefaults)
+    {
+        $this->systemDefaults = $systemDefaults;
     }
 }